Fix 64bit warnings

+++ OAProblem.m
@@ -109,7 +109,7 @@ NSString *token_not_renewable = @"token_not_renewable";
 	return NO;
 }
 
-- (int)code {
+- (NSInteger)code {
 	return [[[self class] validProblems] indexOfObject:problem];
 }
